+++ src/components/table/TableBilling.tsx
@@ -4,9 +4,9 @@ import { Link } from 'react-router-dom';
 import SearchBar from '../searchbar/SearchBar';
 import './Table.css';
 import { getPersonById } from '../../api/PersonAPI';
-import  getClientForPerson  from '../../pages/staffer/functions/forPostulates/getClientForPerson';
-import getProjectForPerson from '../../pages/staffer/functions/forPostulates/getProjectForPerson';
-import getBillRateForBilling from '../../pages/staffer/functions/forPostulates/getBillRateForBilling';
+import  getClientForPerson  from '../../pages/resourceManager/functions/getClientForPerson';
+import getProjectForPerson from '../../pages/resourceManager/functions/getProjectForPerson';
+import getBillRateForBilling from '../../pages/resourceManager/functions/getBillRateForBilling';
 
 interface Props {
   entity: (Employee)[];
